switched from int to uint_8 for rgb variables

+++ ffmpeg.h
@@ -1,8 +1,8 @@
 
 typedef struct rgbPixel {
-    int r;
-    int g;
-    int b;
+    uint8_t r;
+    uint8_t g;
+    uint8_t b;
 } RGBPixel;
 
 typedef struct frame {
